Move alert create to Utils class
